Move to ws folder for rosdep command, fixes #876
Executing the rosdep command from the root of repository (!= workspace),
results in multiple packages found with the same name when run from the
CI on GitLab.

Unable to reproduce this from inside the same container as the CI runs,
but this fix seems to circumvent the issue by restricting the folders
rosdep can search in. Functionally this should behave the same.